# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateUserRequest:
    properties:
      email:
        example: joao@empresa.com
        type: string
      name:
        example: João Silva
        type: string
      password:
        example: senha123
        type: string
      role:
        example: user
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  handlers.GenerateAPIKeyRequest:
    properties:
      name:
        example: Minha API Key
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: joao@empresa.com
        type: string
      password:
        example: senha123
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact:
    email: support@ventros.com
    name: API Support
  description: API para gerenciamento de CRM com eventos e workflows
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ventros CRM API
  version: "1.0"
paths:
...
  /api/v1/auth/api-key:
    post:
      consumes:
      - application/json
      description: Gera uma nova API key para o usuário
      parameters:
      - description: API key request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GenerateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: API key generated
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Gerar API Key
      tags:
      - AUTH - Authentication
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Autentica um usuário e retorna API key
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication failed
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - AUTH - Authentication
  /api/v1/auth/profile:
    get:
      description: Obtém o perfil do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obter Perfil
      tags:
      - AUTH - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário no sistema com projeto e pipeline default
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Criar Usuário
      tags:
      - AUTH - Authentication
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key authentication for service-to-service communication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"